d64_reader/writer merged into d64, fix in disks.s
