--- firmware/d64_writer.h
@@ -1,31 +1,0 @@
-/*
- * Copyright (c) 2019-2020 Kim Jørgensen, Sandor Vass
- *
- * This software is provided 'as-is', without any express or implied
- * warranty.  In no event will the authors be held liable for any damages
- * arising from the use of this software.
- *
- * Permission is granted to anyone to use this software for any purpose,
- * including commercial applications, and to alter it and redistribute it
- * freely, subject to the following restrictions:
- *
- * 1. The origin of this software must not be misrepresented; you must not
- *    claim that you wrote the original software. If you use this software
- *    in a product, an acknowledgment in the product documentation would be
- *    appreciated but is not required.
- * 2. Altered source versions must be plainly marked as such, and must not be
- *    misrepresented as being the original software.
- * 3. This notice may not be removed or altered from any source distribution.
- *
- ******************************************************************************
- */
-
-static const uint8_t d64_sector_num[40] =
-{
-    21, 21, 21, 21, 21, 21, 21,     
-    21, 21, 21, 21, 21, 21, 21, 21,  // 1-17: 21s, 18-24: 19s, 25-30: 18s, 31-40: 17s
-    21, 21, 19, 19, 19, 19, 19, 19,
-    19, 18, 18, 18, 18, 18, 18, 17,
-    17, 17, 17, 17, 17, 17, 17, 17,
-    17
-};
